AsyncResultCallback class moved to InvocationBuilder
